need to do norm

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: ngaurama <ngaurama@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/02 09:01:07 by ngaurama          #+#    #+#             */
-/*   Updated: 2024/11/05 19:49:50 by ngaurama         ###   ########.fr       */
+/*   Updated: 2024/11/07 13:38:50 by ngaurama         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ int ft_isdigit(int c);
 int ft_isalnum(int c);
 int ft_isascii(int c);
 int ft_isprint(int c);
-int ft_strlen(char const *s);
+size_t ft_strlen(char const *s);
 void *ft_memset(void *b, int c, size_t len);
 void ft_bzero(void *s, size_t n);
 void *ft_memcpy(void *dest, const void *src, size_t n);
